+++ CHANGELOG.md
@@ -7,7 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.21.0] - 2023-08-12
 ### Added
-- [@mjkuranda](https://github.com/mjkuranda): GameBar component.
+- [@mjkuranda](https://github.com/mjkuranda): `GameBar` component.
+- [@mjkuranda](https://github.com/mjkuranda): `LongGameBar` component.
+- [@mjkuranda](https://github.com/mjkuranda): Time and durability as game bars.
+- [@mjkuranda](https://github.com/mjkuranda): Score, nickname and missile count as text on the screen.
+- [@mjkuranda](https://github.com/mjkuranda): Empty input for your name gives `Noname` player name.
 
 ## [0.20.0] - 2023-08-11
 ### Added